Refactor test class
