+++ corekit/src/test/java/com/midtrans/sdk/corekit/models/merchant/checkout/request/optional/BillInfoModelTest.java
@@ -1,6 +1,6 @@
 package com.midtrans.sdk.corekit.models.merchant.checkout.request.optional;
 
-import com.midtrans.sdk.corekit.core.merchant.model.checkout.request.optional.BillInfoModel;
+import com.midtrans.sdk.corekit.core.api.merchant.model.checkout.request.optional.BillInfoModel;
 
 import org.junit.Before;
 import org.junit.Test;
