+++ corekit/src/test/java/com/midtrans/sdk/corekit/models/merchant/checkout/request/optional/customer/CustomerDetailsTest.java
@@ -1,6 +1,6 @@
 package com.midtrans.sdk.corekit.models.merchant.checkout.request.optional.customer;
 
-import com.midtrans.sdk.corekit.core.merchant.model.checkout.request.optional.customer.CustomerDetails;
+import com.midtrans.sdk.corekit.core.api.merchant.model.checkout.request.optional.customer.CustomerDetails;
 
 import org.junit.Before;
 import org.junit.Test;
