+++ corekit/src/test/java/com/midtrans/sdk/corekit/models/merchant/checkout/request/specific/banktransfer/BcaBankFreeTextLanguageTest.java
@@ -1,6 +1,6 @@
 package com.midtrans.sdk.corekit.models.merchant.checkout.request.specific.banktransfer;
 
-import com.midtrans.sdk.corekit.core.merchant.model.checkout.request.specific.banktransfer.BcaBankFreeTextLanguage;
+import com.midtrans.sdk.corekit.core.api.merchant.model.checkout.request.specific.banktransfer.BcaBankFreeTextLanguage;
 
 import org.junit.Before;
 import org.junit.Test;
