+++ corekit/src/test/java/com/midtrans/sdk/corekit/models/merchant/checkout/request/specific/creditcard/InstallmentTest.java
@@ -1,6 +1,6 @@
 package com.midtrans.sdk.corekit.models.merchant.checkout.request.specific.creditcard;
 
-import com.midtrans.sdk.corekit.core.merchant.model.checkout.request.specific.creditcard.SavedToken;
+import com.midtrans.sdk.corekit.core.api.merchant.model.checkout.request.specific.creditcard.SavedToken;
 
 import org.junit.Before;
 import org.junit.Test;
