+++ corekit/src/test/java/com/midtrans/sdk/corekit/models/snap/transaction/TransactionOptionResponseTest.java
@@ -1,6 +1,6 @@
 package com.midtrans.sdk.corekit.models.snap.transaction;
 
-import com.midtrans.sdk.corekit.core.snap.model.transaction.response.PaymentInfoResponse;
+import com.midtrans.sdk.corekit.core.api.snap.model.paymentinfo.PaymentInfoResponse;
 
 import org.junit.Before;
 import org.junit.Test;
