+++ corekit/src/test/java/com/midtrans/sdk/corekit/models/snap/transaction/merchantdata/MerchantPreferencesTest.java
@@ -1,6 +1,6 @@
 package com.midtrans.sdk.corekit.models.snap.transaction.merchantdata;
 
-import com.midtrans.sdk.corekit.core.snap.model.transaction.response.merchantdata.MerchantPreferences;
+import com.midtrans.sdk.corekit.core.api.snap.model.paymentinfo.merchantdata.MerchantPreferences;
 
 import org.junit.Before;
 import org.junit.Test;
