+++ corekit/src/test/java/com/midtrans/sdk/corekit/models/snap/transaction/promo/PromoResponseTest.java
@@ -1,6 +1,6 @@
 package com.midtrans.sdk.corekit.models.snap.transaction.promo;
 
-import com.midtrans.sdk.corekit.core.snap.model.transaction.response.promo.PromoResponse;
+import com.midtrans.sdk.corekit.core.api.snap.model.paymentinfo.promo.PromoResponse;
 
 import org.junit.Before;
 import org.junit.Test;
