+++ corekit/src/test/java/com/midtrans/sdk/corekit/models/snap/transaction/promo/PromoTest.java
@@ -1,6 +1,6 @@
 package com.midtrans.sdk.corekit.models.snap.transaction.promo;
 
-import com.midtrans.sdk.corekit.core.snap.model.transaction.response.promo.Promo;
+import com.midtrans.sdk.corekit.core.api.snap.model.paymentinfo.promo.Promo;
 
 import org.junit.Before;
 import org.junit.Test;
